Revert "Fix tests"
