--- generators/app/templates/_schema.yml
@@ -1,4 +1,0 @@
-#
-# Please see byline docs for more informaiton
-# https://github.com/nymag/byline#defining-data-for-your-component-schemayml
-#
