--- generators/app/templates/_server.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = function (url, settings, data) {
-  
-};
