--- generators/context/templates/_all.css
@@ -1,7 +1,0 @@
-.<%= componentName %>{
-  <% if(typeof desktopOnlyDisplay == "undefined"){ %>
-    display: none;
-   <% } else{ %>
-    position: relative;
-  <% } %>
-}
